+++ addOns/ascanrulesBeta/src/main/java/org/zaproxy/zap/extension/ascanrulesBeta/InsecureHttpMethodScanRule.java
@@ -580,6 +580,11 @@ public class InsecureHttpMethodScanRule extends AbstractAppPlugin {
         LOGGER.debug("Request Method: {}", httpMethod);
         LOGGER.debug("Response Code: {}", responseCode);
 
+        if ((httpMethod.equals(HttpRequestHeader.PUT) || httpMethod.equals(HttpRequestHeader.PATCH))
+                && (msg.getResponseHeader().isJson() || msg.getResponseHeader().isXml())) {
+            return;
+        }
+
         if (isPage200(msg) || responseCode == HttpStatusCode.CREATED) {
             evidence =
                     Constant.messages.getString(